Move used components outside of parked
- Improve styling of Segment Input
