+++ src/components/Switch/SwitchIcon/index.tsx
