+++ src/components/Switch/SwitchIcon/styles.module.css
