+++ src/views/ThreeWProjectForm/index.tsx
@@ -38,8 +38,8 @@ import NumberInput from '#components/NumberInput';
 import NonFieldError from '#components/NonFieldError';
 import MultiSelectInput from '#components/MultiSelectInput';
 import useAlert from '#hooks/useAlert';
-import Switch from '#components/parked/Switch';
-import RichTextArea from '#components/parked/RichTextArea';
+import Switch from '#components/Switch';
+import RichTextArea from '#components/RichTextArea';
 import RadioInput from '#components/RadioInput';
 import Checkbox from '#components/Checkbox';
 import TextOutput from '#components/TextOutput';
